refactor: move conversion of UserModulePair to its own adapter
