+++ Sweechat/Models/FirebaseUserModulePair/FirebaseUserModulePairQuery.swift
@@ -15,7 +15,7 @@ class FirebaseUserModulePairQuery {
                     os_log("Error listening for channel updates: \(error?.localizedDescription ?? "No error")")
                     return
                 }
-                let pair: FirebaseUserModulePair? = FirebaseUserModulePairFacade.convert(document: document)
+                let pair: FirebaseUserModulePair? = FirebaseUserModulePairAdapter.convert(document: document)
                 onCompletion(pair)
             }
     }
